+++ lib/Gitlab/Api/IssueLinks.php
@@ -16,30 +16,38 @@ class IssueLinks extends AbstractApi
     }
 
     /**
-     * @param int|string $source_project_id
-     * @param int|string $source_issue_iid
+     * @param int|string $project_id
+     * @param int        $issue_iid
      * @param int|string $target_project_id
-     * @param int|string $target_issue_iid
+     * @param int        $target_issue_iid
+     * @param array      $parameters        {
+     *
+     *     @var string $link_type
+     * }
      *
      * @return mixed
      */
-    public function create($source_project_id, $source_issue_iid, $target_project_id, $target_issue_iid)
+    public function create($project_id, $issue_iid, $target_project_id, $target_issue_iid, array $parameters = [])
     {
-        return $this->post($this->getProjectPath($source_project_id, 'issues/'.$this->encodePath($source_issue_iid).'/links'), [
-            'target_project_id' => $target_project_id,
-            'target_issue_iid' => $target_issue_iid,
-        ]);
+        $parameters['target_project_id'] = $target_project_id;
+        $parameters['target_issue_iid'] = $target_issue_iid;
+
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/links'), $parameters);
     }
 
     /**
      * @param int|string $project_id
      * @param int        $issue_iid
-     * @param int        $issue_link_id
+     * @param int|string $issue_link_id
+     * @param array      $parameters        {
+     *
+     *     @var string $link_type
+     * }
      *
      * @return mixed
      */
-    public function remove($project_id, $issue_iid, $issue_link_id)
+    public function remove($project_id, $issue_iid, $issue_link_id, array $parameters = [])
     {
-        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)).'/links/'.$this->encodePath($issue_link_id));
+        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)).'/links/'.$this->encodePath($issue_link_id), $parameters);
     }
 }